fix: update troubleshooting URL format and improve charge current limit calculation

+++ custom_components/chargepoint_peblar/number.py
@@ -101,7 +101,7 @@ class ChargePointCurrentLimitNumber(ChargePointEntity, NumberEntity):
                 self._attr_native_min_value,
                 self._attr_native_max_value,
             )
-        ma = int(round(clamped * 1000))
+        ma = round(clamped * 1000)
         try:
             await self.coordinator.client.async_patch_evinterface(
                 {"ChargeCurrentLimit": ma}
